Move fixtures into a proper fixture
